+++ controllers/api/commentRoutes.js
@@ -19,3 +19,5 @@ router.post('/', withAuth, async (req, res) => {
       res.status(500).json(err);
     }
 });
+
+module.exports = router;
