+++ controllers/api/postRoutes.js
@@ -63,3 +63,5 @@ router.delete('/:id', withAuth, async (req, res) => {
       res.status(500).json(err);
     }
 });
+
+module.exports = router;
