+++ controllers/index.js
@@ -1,7 +1,7 @@
 const router = require('express').Router();
 
 const apiRoutes = require('./api');
-const homeRoutes = require('./homeRoutes.js');
+const homeRoutes = require('./homepageRoutes');
 
 router.use('/', homeRoutes);
 router.use('/api', apiRoutes);
