+++ models/User.js
@@ -1,5 +1,6 @@
 const { Model, DataTypes } = require('sequelize');
 const sequelize = require('../config/connections');
+const bcrypt = require('bcrypt');
 
 class User extends Model {}
 
@@ -26,16 +27,9 @@ User.init(
   },
   {
     hooks: {
-      // Use the beforeCreate hook to work with data before a new instance is created
       beforeCreate: async (newUserData) => {
-        // In this case, we are taking the user's email address, and making all letters lower case before adding it to the database.
-        newUserData.username = await newUserData.username.toLowerCase();
+        newUserData.password = await bcrypt.hash(newUserData.password, 10);
         return newUserData;
-      },
-      // Here, we use the beforeUpdate hook to make all of the characters lower case in an updated email address, before updating the database.
-      beforeUpdate: async (updatedUserData) => {
-        updatedUserData.username = await updatedUserData.username.toLowerCase();
-        return updatedUserData;
       },
     },
     sequelize,
