+++ public/js/homepage.js
@@ -1,0 +1,1 @@
+//TODO display all posts that were done by all users
