+++ seeds/userSeeds.json
@@ -1,18 +1,18 @@
 [
     {
       "username": "Sal",
-      "password": "pass123"
+      "password": "pass1230"
     },
     {
-      "username": "Sal",
-      "password": "pass456"
+      "username": "Val",
+      "password": "pass4560"
     },
     {
-      "username": "Sal",
-      "password": "pass789"
+      "username": "Jack",
+      "password": "pass7890"
     },
     {
-      "username": "Sal",
-      "password": "pass147"
+      "username": "Mat",
+      "password": "pass1470"
     }
 ]
