+++ views/dashboard.handlebars
@@ -1,0 +1,11 @@
+<h2>Your Dashboard</h2>
+<div>
+    <div>
+
+    </div>
+</div>
+<div>
+    <button id="add-newpost" type="button">+ New Post</button>
+</div> 
+
+<script src="./../public/js/dashboard.js"></script>
