+++ views/homepage.handlebars
@@ -1,0 +1,9 @@
+<h2>The Tech Blog</h2>
+<div>
+    <div>
+
+    </div>
+</div>
+  
+
+<script src="./../public/js/homepage.js"></script>
