+++ views/newpost.handlebars
@@ -1,0 +1,25 @@
+<h2>Your Dashboard</h2>
+<div>
+    <div>
+        <h3>Create New Post</h3>
+    </div>
+    <form id="newpost-form">
+        <div>
+            <label for="newpost-title">Title:</label>
+            <input type="input" id="newpost-title" />
+        </div>
+        <div>
+            <label for="newpost-content">password:</label>
+            <input type="input" id="newpost-content" />
+        </div>
+        <div>
+            <button type="submit">Create</button>
+        </div>
+    </form>
+    <div>
+        <button id="add-newpost" type="button">+ New Post</button>
+    </div>
+</div>
+  
+
+<script src="./../public/js/homepage.js"></script>
