+++ views/signin.handlebars
@@ -17,4 +17,4 @@
     <button id="change-to-signup" type="button">Sign up instead</button>
   </div>
 
-<script src="./../public/js/login.js"></script>
+<script src="/js/login.js"></script>
