+++ views/signup.handlebars
@@ -17,4 +17,4 @@
     <button id="change-to-login" type="submit">Login instead</button>
   </div>
 
-<script src="./../public/js/login.js"></script>
+<script src="./../public/js/signup.js"></script>
